Updating dates for CGT to provide a more realistic resolution deadline.
CGT is a 3 part service of, which is just 1 of many services for DDCT Live Services and is currently in the process of being upgraded along with many other services. This  date changes allows for a more realistic timeline before getting the service retested after the current ui upgrade is complete.