Statement updates ahead of March reporting run
Statements to be archived:
/integration-hub
This statement is for internal aspect of API hub and MWS team have the actual statement for this aspect in their statements - this one has been agreed to be archived as not needed.

/change-of-circumstances
accessibility.team inbox notified service is to be decommissioned 13-11-2023.

/enquire-plastic-packaging-tax
/register-for-plastic-packaging-tax
Statement creation and last updated dates in the future. Changed to reflect date these were created.